requirements.txt
asgiref==3.7.2
cachetools==5.3.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.2
cssselect==1.2.0
cssutils==2.9.0
defusedxml==0.7.1
Django==4.2.9
django-allauth==0.60.1
django-environ==0.11.2
idna==3.6
instamojo-wrapper==1.2.0
lxml==5.1.0
oauthlib==3.2.2
premailer==3.10.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
pypng==0.20220715.0
python3-openid==3.2.0
qrcode==7.4.2
requests==2.31.0
requests-oauthlib==1.3.1
sqlparse==0.4.4
typing_extensions==4.9.0
urllib3==2.2.0
yagmail==0.15.293
